Move all util methods into a separate module
This commit moves all util methods that have previously been in the test class into a separate module. This module is built into a shadow jar which needs to be added into the user's project dependencies in order for our generated tests to be able to use its functionality.
