+++ utbot-codegen-utils/build.gradle
@@ -1,0 +1,15 @@
+plugins {
+    id "com.github.johnrengelman.shadow" version "6.1.0"
+}
+
+apply from: "${parent.projectDir}/gradle/include/jvm-project.gradle"
+
+shadowJar {
+    configurations = [project.configurations.compileClasspath]
+    archiveClassifier.set('')
+    minimize()
+}
+
+dependencies {
+    compileOnly group: 'org.mockito', name: 'mockito-core', version: mockito_version
+}
