+++ utbot-framework/build.gradle
@@ -15,6 +15,7 @@ dependencies {
     api project(':utbot-core')
     api project(':utbot-instrumentation')
     api project(':utbot-summary')
+    api project(':utbot-codegen-utils')
     implementation 'junit:junit:4.13.1'
     api project(':utbot-framework-api')
