+++ utbot-framework/src/main/kotlin/org/utbot/framework/codegen/model/constructor/tree/CgFieldStateManager.kt
@@ -2,6 +2,7 @@ package org.utbot.framework.codegen.model.constructor.tree
 
 import org.utbot.framework.codegen.model.constructor.builtin.forName
 import org.utbot.framework.codegen.model.constructor.builtin.getArrayElement
+import org.utbot.framework.codegen.model.constructor.builtin.utUtilsClassId
 import org.utbot.framework.codegen.model.constructor.context.CgContext
 import org.utbot.framework.codegen.model.constructor.context.CgContextOwner
 import org.utbot.framework.codegen.model.constructor.util.CgComponents
@@ -219,7 +220,7 @@ internal class CgFieldStateManagerImpl(val context: CgContext)
             val expression = when (val newElement = path[index++]) {
                 is FieldAccess -> {
                     val field = newElement.field
-                    testClassThisInstance[getFieldValue](prev, stringLiteral(field.name))
+                    utUtilsClassId[getFieldValue](prev, stringLiteral(field.name))
                 }
                 is ArrayElementAccess -> {
                     Array::class.id[getArrayElement](prev, newElement.index)
@@ -244,7 +245,7 @@ internal class CgFieldStateManagerImpl(val context: CgContext)
             } else {
                 newVar(classCgClassId) { Class::class.id[forName](owner.name) }
             }
-            newVar(objectClassId) { testClassThisInstance[getStaticFieldValue](ownerClass, stringLiteral(firstField.name)) }
+            newVar(objectClassId) { utUtilsClassId[getStaticFieldValue](ownerClass, stringLiteral(firstField.name)) }
         }
         val path = fieldPath.elements
         val remainingPath = fieldPath.copy(elements = path.drop(1))
